Desktop/AGRISCAN-MODEL-main/backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
aiofiles
numpy
tensorflow
python-jose
python-dotenv
pydantic
sqlalchemy
